+++ aioeventlet.py
@@ -28,6 +28,12 @@ if eventlet.patcher.is_monkey_patched('socket'):
     # Examples: socket.socket(), socket.socketpair(),
     # threading.current_thread().
     asyncio.base_events.socket = socket
+    asyncio.base_events.threading = threading
+    if hasattr(threading, 'get_ident'):
+        asyncio.base_events._get_thread_ident = threading.get_ident
+    else:
+        # Python 2
+        asyncio.base_events._get_thread_ident = threading._get_ident
     asyncio.events.threading = threading
     if sys.platform == 'win32':
         asyncio.windows_events.socket = socket
@@ -315,7 +321,7 @@ def yield_future(future, loop=None):
     Return the result or raise the exception of the future.
 
     The function must not be called from the greenthread
-    of the aioeventlet event loop.
+    running the aioeventlet event loop.
     """
     future = asyncio.async(future, loop=loop)
     if future._loop._greenthread == eventlet.getcurrent():