+++ tests/__init__.py
@@ -1,13 +1,10 @@
 import aioeventlet
+import sys
 try:
     import asyncio
 except ImportError:
     import trollius as asyncio
-try:
-    # On Python 2.6, unittest2 is needed to get new features like addCleanup()
-    import unittest2 as unittest
-except ImportError:
-    import unittest
+import unittest
 try:
     from unittest import mock
 except ImportError:
@@ -22,3 +19,7 @@ class TestCase(unittest.TestCase):
         self.loop = policy.get_event_loop()
         self.addCleanup(self.loop.close)
         self.addCleanup(asyncio.set_event_loop, None)
+
+    if sys.version_info < (3,):
+        def assertRaisesRegex(self, *args):
+            return self.assertRaisesRegexp(*args)