+++ doc/changelog.rst
@@ -1,11 +1,24 @@
 Changelog
 =========
 
+2016-02-22: Version 0.5
+-----------------------
+
+* Unit tests now use the aiotest library.
+* Fix for eventlet used with monkey-patching: inject the original threading
+  module and the original threading.get_ident() function in
+  asyncio.base_events.
+* Drop support for Python 2.6 and Python 3.2. aioeventlet depends on trollius
+  and pip which don't support these Python versions anymore.
+
 2014-12-03: Version 0.4
 -----------------------
 
+* Add run_aiotest.py
+* tox now also run the aiotest test suite
 * Rename the project from ``aiogreen`` to ``aioeventlet``
 * Rename the ``link_future()`` function to :func:`yield_future`
+* Running tests with eventlet monkey-patching now works with Python 3.
 
 2014-11-23: version 0.3
 -----------------------